V03 - Enhanced the OCR pre- and postprocesses
The tags, which precede each paragraph, have been reworked to contain a
repetitive numerical pattern. This enhances the ability of the OCR
post-process to link a paragraph tag to its contents when extracting the
OCR’es texts from the OCRResult.txt file.
The horizontal alignment on the Tesseract input image of the various
lines within a single text bubble has been improved.
The output of the Tesseract OCR process is “scrubbed” prior to being
loaded into the cbrTekStraktor application. The scrubbing process is
rather coarse and removes all characters outside the 0x00000020 and
0x000000ff range.
The contents of the Language File are overwritten after each OCR run